refactor: integrate response of kakao map's keyword search api
