--- backend/src/main/java/com/example/backend/domain/restaurant/api/response/RestaurantDocument.java
@@ -1,18 +1,0 @@
-package com.example.backend.domain.restaurant.api.response;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record RestaurantDocument(
-        String id,
-        @JsonProperty("place_name") String placeName,
-        @JsonProperty("category_name") String categoryName,
-        @JsonProperty("category_group_code") String categoryGroupCode,
-        @JsonProperty("category_group_name") String categoryGroupName,
-        String phone,
-        @JsonProperty("address_name") String addressName,
-        @JsonProperty("road_address_name") String roadAddressName,
-        String x,
-        String y,
-        @JsonProperty("place_url") String placeUrl,
-        String distance
-) {}
