--- backend/src/main/java/com/example/backend/domain/restaurant/api/response/RestaurantSearchResponse.java
@@ -1,8 +1,0 @@
-package com.example.backend.domain.restaurant.api.response;
-
-import java.util.List;
-
-public record RestaurantSearchResponse(
-        Meta meta,
-        List<RestaurantDocument> documents
-) {}
