--- backend/src/main/java/com/example/backend/domain/restaurant/api/response/SameName.java
@@ -1,11 +1,0 @@
-package com.example.backend.domain.restaurant.api.response;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import java.util.List;
-
-public record SameName(
-        List<String> region,
-        String keyword,
-        @JsonProperty("selected_region") String selectedRegion
-) {}
