+++ backend/src/main/java/com/example/backend/domain/restaurant/service/RestaurantService.java
@@ -1,7 +1,7 @@
 package com.example.backend.domain.restaurant.service;
 
+import com.example.backend.domain.restaurant.api.response.KakaoPlacesSearchResponse;
 import com.example.backend.domain.restaurant.api.response.RestaurantResponse;
-import com.example.backend.domain.restaurant.api.response.RestaurantSearchResponse;
 import com.example.backend.domain.restaurant.api.response.RestaurantsResponse;
 import com.example.backend.domain.restaurant.client.KakaoSearchApiClient;
 import com.example.backend.domain.restaurant.entity.Category;
@@ -33,7 +33,7 @@ public class RestaurantService {
     private final CategoryRepository categoryRepository;
     private final KakaoSearchApiClient kakaoSearchApiClient;
 
-    public RestaurantSearchResponse search(RestaurantSearchRequest restaurantSearchRequest) {
+    public KakaoPlacesSearchResponse search(RestaurantSearchRequest restaurantSearchRequest) {
         return kakaoSearchApiClient.searchRestaurants(restaurantSearchRequest);
     }
