+++ backend/src/main/java/com/example/backend/domain/restaurant/service/response/RestaurantCreateRequest.java
@@ -1,7 +1,5 @@
 package com.example.backend.domain.restaurant.service.response;
 
-import com.example.backend.domain.restaurant.api.response.RestaurantDocument;
-
 
 public record RestaurantCreateRequest(
         Long categoryId,
@@ -11,13 +9,13 @@ public record RestaurantCreateRequest(
         String latitude
 ) {
 
-    public static RestaurantCreateRequest of(RestaurantDocument doc, Long categoryId) {
+    public static RestaurantCreateRequest of(Long categoryId, String placeName, String roadAddressName, String longitude, String latitude) {
         return new RestaurantCreateRequest(
                 categoryId,
-                doc.placeName(),
-                doc.addressName(),
-                doc.x(),
-                doc.y()
+                placeName,
+                roadAddressName,
+                longitude,
+                latitude
         );
     }
 }
